로그인 권한

+++ src/main/java/com/example/spring/demo/controller/UsrArticleController.java
@@ -2,13 +2,17 @@ package com.example.spring.demo.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.example.spring.demo.service.ArticleService;
+import com.example.spring.demo.util.Ut;
 import com.example.spring.demo.vo.Article;
+import com.example.spring.demo.vo.ResultData;
 
 @Controller
 public class UsrArticleController {
@@ -18,56 +22,99 @@ public class UsrArticleController {
 	// 액션 매서드 시작
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
-	public Article doAdd(String title, String body) {
-		int id = articleService.writeArticle(title, body);
+	public ResultData<Article> doAdd(HttpSession httpSession, String title, String body) {
+		// 로그인 후 게시글 작성 권한
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
+		// 종료
+		// 게시글 작성
+		if (Ut.empty(title)) {
+			return ResultData.from("F-1", "title(을)를 입력해주세요.");
+		}
+
+		if (Ut.empty(body)) {
+			return ResultData.from("F-2", "body(을)를 입력해주세요.");
+		}
+		// 종료
+		// S-1
+		// 4번 게시물이 생성되었습니다.
+		ResultData<Integer> writeArticleRd = articleService.writeArticle(loginedMemberId, title, body);
+		int id = writeArticleRd.getData1();
 
 		Article article = articleService.getArticle(id);
-		return article;
+		return ResultData.newData(writeArticleRd, article);
 	}
 
 	@RequestMapping("/usr/article/getArticles")
 	@ResponseBody
-	public List<Article> getArticles() {
-		return articleService.getArticles();
+	public ResultData<List<Article>> getArticles() {
+		List<Article> articles = articleService.getArticles();
+		return ResultData.from("S-1", "게시물 리스트 입니다.", articles);
 	}
 
 	@RequestMapping("/usr/article/getArticle")
 	@ResponseBody
-	public Object getArticle(int id) {
+	public ResultData<Article> getArticle(int id) {
 		Article article = articleService.getArticle(id);
 		if (article == null) {
-			return id + "번 게시물이 존재하지 않습니다.";
+			return ResultData.from("F-1", Ut.f("%d번 게시글이 존재하지 않습니다.", id));
 		}
 
-		return article;
+		return ResultData.from("S-1", Ut.f("%d번 게시글입니다.", id), article);
 	}
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public String doDelete(int id) {
-		Article article = articleService.getArticle(id);
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+		// 로그인 후 게시글 작성 권한
+		boolean isLogined = false;
+		int loginedMemberId = 0;
 
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		} 
+
+		Article article = articleService.getArticle(id);
+		
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", "삭제 권한이 없습니다.");
+		} 
+		
 		if (article == null) {
-			return id + "번 게시물이 존재하지 않습니다.";
+			return ResultData.from("F-1", Ut.f("%d번 게시글이 존재하지 않습니다.", id));
 		}
 
 		articleService.deleteArticle(id);
 
-		return id + "번 게시물을 삭제하였습니다.";
+		return ResultData.from("S-1", Ut.f("%d번 게시글이 삭제되었습니다.", id), id);
 	}
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public String doModify(int id, String title, String body) {
+	public ResultData<Integer> doModify(int id, String title, String body) {
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
-			return id + "번 게시물이 존재하지 않습니다.";
+			return ResultData.from("F-1", Ut.f("%d번 게시글이 존재하지 않습니다.", id));
 		}
 
 		articleService.modifyArticle(id, title, body);
 
-		return id + "번 게시물을 수정하였습니다.";
+		return ResultData.from("S-1", Ut.f("%d번 게시글이 수정되었니다.", id), id);
 	}
 
 	// 액션 매서드 끝
